wip: Rename `utils` to `shared/lib`
