+++ src/families/base.jsx
@@ -4,11 +4,11 @@ import { Component } from "react";
 
 import { Tooltip } from "../shared/ui/components";
 import { TextInput, TextInputWithUnits } from "../widgets/editor/elements";
-import { ArgsAccount, ArgsBig, ArgsError, ArgsJSON, ArgsString } from "../utils/args";
-import Call from "../utils/call";
-import { toGas, toYocto, formatTokenAmount, unitToDecimals } from "../utils/converter";
-import { errorMsg } from "../utils/errors";
-import { hasContract } from "../utils/contracts/generic";
+import { ArgsAccount, ArgsBig, ArgsError, ArgsJSON, ArgsString } from "../shared/lib/args";
+import Call from "../shared/lib/call";
+import { toGas, toYocto, formatTokenAmount, unitToDecimals } from "../shared/lib/converter";
+import { errorMsg } from "../shared/lib/errors";
+import { hasContract } from "../shared/lib/contracts/generic";
 import "./base.scss";
 
 export class BaseTask extends Component {
