+++ src/families/near/transfer.jsx
@@ -1,11 +1,11 @@
 import { InputAdornment } from "@mui/material";
 import React from "react";
 import { TextInput, TextInputWithUnits } from "../../widgets/editor/elements";
-import { ArgsAccount, ArgsBig, ArgsError, ArgsObject, ArgsString } from "../../utils/args";
-import Call from "../../utils/call";
-import { toGas, formatTokenAmount, unitToDecimals } from "../../utils/converter";
-import { view } from "../../utils/wallet";
-import { errorMsg } from "../../utils/errors";
+import { ArgsAccount, ArgsBig, ArgsError, ArgsObject, ArgsString } from "../../shared/lib/args";
+import Call from "../../shared/lib/call";
+import { toGas, formatTokenAmount, unitToDecimals } from "../../shared/lib/converter";
+import { view } from "../../shared/lib/wallet";
+import { errorMsg } from "../../shared/lib/errors";
 import { BaseTask } from "../base";
 import debounce from "lodash.debounce";
 import "./near.scss";
