+++ src/shared/lib/contracts/generic.ts
@@ -1,5 +1,4 @@
-import { viewAccount } from "../wallet"
-
+import { viewAccount } from "../wallet";
 
 /**
  * check if there's a contract deployed on given NEAR address.
@@ -7,10 +6,10 @@ import { viewAccount } from "../wallet"
  *
  * @param {string} address
  */
-async function hasContract (address: string): Promise<boolean> {
-  const accountInfo = await viewAccount(address);
-  const codeHash: string = accountInfo.code_hash;
-  return codeHash !== "11111111111111111111111111111111";
+async function hasContract(address: string): Promise<boolean> {
+    const accountInfo = await viewAccount(address);
+    const codeHash: string = accountInfo.code_hash;
+    return codeHash !== "11111111111111111111111111111111";
 }
 
 // TODO: method to list all available functions from a contract
