+++ src/shared/lib/contracts/multicall.ts
