+++ src/shared/lib/errors.ts
