+++ src/shared/lib/loader.ts
