+++ src/shared/lib/standards/fungibleToken.ts
