+++ src/shared/lib/standards/storageManagement.ts
